# Movie-client/src/api-client/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MovieReview API Specification",
    "description": "OpenApi documentation for MovieReview API",
    "termsOfService": "https://moviereview.com/terms",
    "contact": {
      "name": "MovieReview Team",
      "url": "https://moviereview.com",
      "email": "contact@moviereview.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    },
    {
      "url": "https://api.moviereview.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/movies/{imdbId}": {
      "get": {
        "tags": ["movie-controller"],
        "operationId": "getMovieByImdbId",
        "parameters": [
          {
            "name": "imdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "-media-type=application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieWithReviewsDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["movie-controller"],
        "operationId": "updateMovie",
        "parameters": [
          {
            "name": "imdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movies"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "-media-type=application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movies"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GrantedAuthority": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string"
          }
        }
      },
      "Movies": {
        "required": ["imdbId", "title"],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "imdbId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "releaseDate": {
            "type": "string"
          },
          "trailerLink": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "backdrops": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reviewIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reviews"
            }
          }
        }
      },
      "ObjectId": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Reviews": {
        "required": ["body"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string"
          },
          "lastModifiedBy": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "movie": {
            "$ref": "#/components/schemas/Movies"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Role": {
        "required": ["createdDate", "lastUpdatedDate"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdatedDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "User": {
        "required": ["createdDate", "email", "lastUpdatedDate", "username"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "accountLocked": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "accountNonLocked": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "MovieWithReviewsDTO": {
        "type": "object",
        "properties": {
          "imdbId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "releaseDate": {
            "type": "string"
          },
          "trailerLink": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "backdrops": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewDTO"
            }
          }
        }
      },
      "ReviewDTO": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT authentication",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}